Rename from entente to lightldapd.
Change all references of "entente" that are not explicitly referring
to the previous entente project to "lightldapd". This changes the
binary, configs, .gitignore, and documentation.

Fix Makefile debian target to not include ".*" hidden files in the
package tarball. This ensures .git and .gitignore are not included in
the debian package source tarball.

Update debian control and changelog files for the new package details.
Fix copyright symlink to point at LICENSE.

Fix README.rst list of included files.

Update NEWS.rst and TODO.rst with these changes.