Improve and document pam.[ch] implementation.
Changed auth_pam() function to take a slightly simpler argument for
the msg argument, requiring that it be a char[PAMMSG_LEN] string. This
is similar to the API used by the nss2pam.[ch] code.

Fixed potential memory leak on pam errors when the pam_conv() handles
more than one msg at a time.