Fix #6 so abandonRequest doesn't close the connection.
This means abandonRequests are ignored and don't close the connection.
They don't cancel the request, since requests will have completed
before we read and process the abandonRequest.

This meets the LDAP spec requirements, which don't require that
abandonRequests do anything. Actually implementing support for
abandoning requests is more complicated and of questionable value.