Fix #11 so Search baseObject works correctly.
This means baseObject can be set to constrain searchs to only users or
groups, and will find results if it matches the end of either the
"ou=persons,..." or "ou=groups,... dn". The baseObject can constrain
the nss scan to only passwd or groups instead of always scanning both.

This also means that Search will no longer report baseObject "invalid"
errors. Searches with objectBase set to a string that doesn't match
the end of the users or groups dn will simply not find anything.