Fix #10 Refactor ldap_request into it's own class.
An ldap_connection can now have multiple ldap_requests in-flight. The
replies for all ldap_requests will be interleaved. This means
abandonRequests can now correctly immediately cancel in-flight
requests, and unbindRequest will immediately cancel all in-flight
requests and close the connection.

An ldap_request has an ldap_response containing a circular dlist of
ldap_reply's with the response messages. This now uses dlist.h for the
dlists.

There is still some tidying up to be done here, but it works.