Don't constrain scope data sources based on attribute names.
This means you must use objectClass or baseObject constraints in
SearchRequests to constrain searches to only passwd or group data.
This makes it easier to add other data sources that might use the same
attribute names in the future.

Remove AttributeDescription_present_scope() method. Make
AttributeValueAssertion_equal_scope() not constrain scope to passwd
for uid and uidNumber attributes.

Fix Filter_scope() to return a full scope for anything except
and/or/not/equalityMatch filters.