Use mbedtls/net_socket.h for binding the server socket.
This is in preparation for using mbedtls for TLS support. Move the
mbedtls_net_bind() operation outside ldap_server_start() and into
main.c so we can setuid() away from root as soon as the port is bound,
and before we initialize the ldap_server. This means
ldap_server_start() now takes an mbedtls_net_context as an argument
for the bound socket which is an attribute of the ldap_server.

Change main.c to use strings for the address and port as used by
mbedtls_net_bind(). Note this also means we can now take hostnames and
are IPv6 compatible, though we still only bind to "127.0.0.1" when
running in loopback only mode.

Update Makefile to use mbedtls.