Use mbedtls_net_* stuff for the connection socket.
Change ldap_server_new() to take and copy a whole mbedtls_net_context
struct argument instead of a pointer, and properly close it in
ldap_server_free(). The mbedtls_net_context struct contains only an fd
int so it is cheap to pass-by-value and copy, and this avoids a bunch
of pointer handling overheads.