Improve anonymous and ssl permissions behaviour.
Make StartTLS require that the server has TLS is enabled (it was
started with -C <certpath>). This fixes a crash bug.

Always allow anonymous binds, and make other binds require TLS if the
sever has TLS enabled. Note the LDAP standard requires supporting
anonymous binds as the way to unbind from a particular user without
dropping the connection.

Only allow anonymous searches if the server has anonymous access
enabled (it was started with -a).

This combination of settings means a server started with TLS enabled
and anonymous disabled will always require encryption and
authentication for searches.

Always cast the uid -1 to (uid_t) to ensure correct type.