Add support for `-U` and `-G` uid/gid ranges.
This means we can server subsets of all the nss uids and gids to fix
https://github.com/dbaarda/LightLdapd/issues/13.

In ldap_server.[ch] add ldap_ranges uids and gids fields to ldap_server and
args to ldap_server_init()

In nss2ldap.c add ldap_ranges uids and gids fields to scope_t and args to
scope_init(). Change SearchRequest_scope() to take an server argument instead
of just a basedn so that it can also get the uids and gids from the server and
set it in the scope. Change scope_(passwd|group)_(iter|next)() methods to skip
over uids/gids that don't match the uids/gids ldap_ranges.

In main.c add support for `-U` and `-G` arguments for setting uids/gids ranges.