Fix `-R chroot` for relative paths and add basic logging.
Change daemon() call to not chdir to '/' so that the chroot() later works with
relative paths, and instead chdir() to '/' explicitly after the chroot(). Move
ldap_server_init() and mbedtls_net_bind() before daemon() so that if these
fail errors get shown on stderr before daemon() has redirected it to /dev/null.

Add basic log.h logging with an API similar to err.h with lerr(), lerrx(),
lwarn(), etc. If DEBUG is defined it uses err.h to send colour output to
stderr, otherwise it uses syslog. Move linit() logging initialization to just
after settings() are parsed so that it is initialized before any errors are
logged.

Change settings_setuid to settings_runuser and move converting it to a runuid
inside main() after logging is initialized, but before switching to the
chroot, so that errors can be logged, and it is resolved outside the chroot.
In settings() use "true" instead of 1 for bool settings, and change the
returncode for bad popt arguments to EX_USAGE.

In ssl.h and utils.h, replace all err(), warn(), etc calls with logging
equivalents.