Change logging to select stderr or syslog based on `-d` daemon mode.
This required making the logging dynamically selectable at runtime instead of
selecting macros at compile-time. This was done by introducing a log_func
pointer to a logging function must take the same args as syslog(). A errlog()
log_func was added that uses vwarnx() for logging if the level is less than or
equal to a log_level value.

The old linit() was replaced by log_init() that calls startlog() if using
syslog, and sets log_func and log_level.

The lerr(), lerrx(), lwarn(), etc macro's were changed call the log_func() via log()
and logx() macro's that add __FILE__, __LINE__, and __FUNCTION__ information.

This unfortunately sacrifices the colored output to stderr, but it's hard to
make it run-time selectable for syslog vs stderr output. I think I can add it
back later.