Add client_id field to ldap_connection and add logging.
This gets and keeps the client_id for incoming connections which is used for
logging.

Connections, requests, and replies are logged with priorities notice, info,
and debug respectively. Events that lead to connections being dropped are
logged at warn level.

This is pretty rudimentary, and uses a nasty hack to get the message type
name, but it does work.