Fix #16: Add signal handlers.
Add SIGHUP, SIGINT, and SIGTERM ev_signal signal handlers. SIGHUP will be used
for reloading configs when/if we ever have them. SIGTERM initiates a clean
shutdown. SIGINT is treated the same as SIGTERM so when running in non-daemon
mode ^C cleanly terminates the server.

Note the main event loop is not killed with ev_break(), only the connection
and signal watchers are stopped, and the ev_loop() will terminate when there
are no more watchers running. This means that existing connections will delay
the shutdown until they complete their requests and disconnect.

This also means a connection that stays open will delay the shutdown
indefinitely, requiring a SIGQUIT or SIGKILL to terminate the process. If this
proves a problem we may need to add a circular list of connections and have
ldap_server_stop() cycle through them with ldap_connection_close() so they all
cleanly shutdown.